requirements.txt
kafka-python==1.4.6
psycopg2==2.8.4
